requirements.txt
requests~=2.31.0
discord.py~=1.7.3
python-dotenv~=1.0.0
openai~=0.11.0
dataset~=1.6.2
DiscordUtils~=1.3.4
